// vhome_mis/src/app/core/api-client-v2/model/mainPagedHandoverMessageResponseMainHandoverMessageRow.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MainPagedHandoverMessageResponseMainHandoverMessageRow { 
    ID: number;
    creator_id: number;
    creator_name: string;
    location_ids: Array<number>;
    location_names: Array<string>;
    message_type_ids: Array<number>;
    message_type_names: Array<string>;
    shift_id: number;
    shift_name: string;
    timestamp: string;
}